# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Nexis License System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/licenses:
    get:
      operationId: listLicenses
      tags: [admin-licenses]
      summary: List all licenses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: planId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Paginated list of licenses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseListResponse"
    post:
      operationId: createLicense
      tags: [admin-licenses]
      summary: Generate a new license
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "201":
          description: Created license
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /admin/licenses/bulk-generate:
    post:
      operationId: bulkGenerateLicenses
      tags: [admin-licenses]
      summary: Bulk generate multiple licenses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkGenerateInput"
      responses:
        "201":
          description: Generated licenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/License"
  /admin/licenses/export:
    get:
      operationId: exportLicenses
      tags: [admin-licenses]
      summary: Export licenses as CSV
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
  /admin/licenses/{id}:
    get:
      operationId: getLicense
      tags: [admin-licenses]
      summary: Get a license by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLicense
      tags: [admin-licenses]
      summary: Update a license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseUpdate"
      responses:
        "200":
          description: Updated license
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
    delete:
      operationId: deleteLicense
      tags: [admin-licenses]
      summary: Revoke and delete a license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/licenses/{id}/suspend:
    post:
      operationId: suspendLicense
      tags: [admin-licenses]
      summary: Suspend a license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /admin/licenses/{id}/activate:
    post:
      operationId: activateLicenseAdmin
      tags: [admin-licenses]
      summary: Re-activate a suspended license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /admin/licenses/{id}/revoke:
    post:
      operationId: revokeLicense
      tags: [admin-licenses]
      summary: Revoke a license permanently
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /admin/licenses/{id}/renew:
    post:
      operationId: renewLicense
      tags: [admin-licenses]
      summary: Renew/extend a license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewInput"
      responses:
        "200":
          description: Renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"

  # ─────────────────────────────────────────────────────────────────
  # Admin — Plans
  # ─────────────────────────────────────────────────────────────────
...
  /admin/devices:
    get:
      operationId: listDevices
      tags: [admin-devices]
      summary: List all registered devices
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: licenseId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Devices list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceListResponse"
  /admin/devices/{id}:
    delete:
      operationId: deleteDevice
      tags: [admin-devices]
      summary: Remove a device binding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─────────────────────────────────────────────────────────────────
  # Admin — Notifications
  # ─────────────────────────────────────────────────────────────────
...
  /admin/activity-logs:
    get:
      operationId: listActivityLogs
      tags: [admin-logs]
      summary: List activity logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: licenseId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLogListResponse"

  # ─────────────────────────────────────────────────────────────────
  # Admin — Settings
  # ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── Auth ──────────────────────────────────────────────────────
...
    License:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        status:
          type: string
          enum: [active, suspended, revoked, expired, trial]
        planId:
          type: ["integer", "null"]
        planName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        maxDevices:
          type: integer
        activatedDevices:
          type: integer
        expiresAt:
          type: ["string", "null"]
        activatedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        notes:
          type: ["string", "null"]
      required: [id, key, status, maxDevices, activatedDevices, createdAt]
    LicenseDetail:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        status:
          type: string
        planId:
          type: ["integer", "null"]
        planName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        maxDevices:
          type: integer
        activatedDevices:
          type: integer
        expiresAt:
          type: ["string", "null"]
        activatedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        notes:
          type: ["string", "null"]
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"
      required: [id, key, status, maxDevices, activatedDevices, createdAt, devices, recentActivity]
    LicenseInput:
      type: object
      properties:
        planId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        maxDevices:
          type: integer
        expiresAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isTrial:
          type: boolean
      required: [maxDevices]
    LicenseUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
        planId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        maxDevices:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    BulkGenerateInput:
      type: object
      properties:
        count:
          type: integer
        planId:
          type: ["integer", "null"]
        maxDevices:
          type: integer
        expiresAt:
          type: ["string", "null"]
        isTrial:
          type: boolean
      required: [count, maxDevices]
    RenewInput:
      type: object
      properties:
        expiresAt:
          type: string
      required: [expiresAt]
    LicenseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/License"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ── Plans ──────────────────────────────────────────────────────
...
    Device:
      type: object
      properties:
        id:
          type: integer
        licenseId:
          type: integer
        deviceId:
          type: string
        deviceInfo:
          type: ["object", "null"]
          additionalProperties: true
        firstSeen:
          type: string
        lastSeen:
          type: string
        isActive:
          type: boolean
      required: [id, licenseId, deviceId, firstSeen, lastSeen, isActive]
    DeviceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ── Notifications ──────────────────────────────────────────────
...
    ActivityLog:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        licenseId:
          type: ["integer", "null"]
        licenseKey:
          type: ["string", "null"]
        deviceId:
          type: ["string", "null"]
        adminId:
          type: ["integer", "null"]
        metadata:
          type: ["object", "null"]
          additionalProperties: true
        createdAt:
          type: string
      required: [id, action, createdAt]
    ActivityLogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ── Settings ─────────────────────────────────────────────────
...