# lib/api-zod/src/generated/types/activityLogListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexis License System API
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityLog } from './activityLog';

export interface ActivityLogListResponse {
  data: ActivityLog[];
  total: number;
  page: number;
  limit: number;
}